Take offset not idx in PER_BUFFER_VALUE_P
This also moves the common "idx == -1" check into PER_BUFFER_VALUE_P.
If idx == -1, it's guaranteed that the corresponding buffer field has
a per-buffer value.  We do this check everywhere we call
PER_BUFFER_VALUE_P, so let's put it in the common code.

This is motivated by later commits which will get rid of idx.