Remove local_flags array in struct buffer
Previously, we used the local_flags array in each struct buffer to
know which field was buffer-local.  Now, we just set the field to
Qunbound if it's not buffer-local, and fall back to the corresponding
field in buffer_defaults.  So we can delete the local_flags array.

Because of this fallback logic, we now no longer need to update all
struct buffer objects whenever we set a value in buffer_defaults,
which provides a substantial performance gain.

The fallback logic applies for all accesses to Lisp_Object fields in
struct buffer, but it should now be very fast: It's one branch on
already-loaded data to check if it's equal to the Qunbound constant.

In these changes, we just add the Qunbound fall back logic, delete
local_flags, and delete several loops which updated all buffers when
buffer_defaults changed.

We also delete some calls to reset_buffer_local_variables on
buffer_defaults and buffer_local_symbols.  Since resetting a
buffer-local now requires setting it to Qunbound, those would set
fields on buffer_defaults and buffer_local_symbols.  Fortunately, the
calls are not necessary anymore.