Don't escape "." in `prin1' unless followed by "?" or nothing
This is a followup to 637dde4aba921435f78d0de769ad74c4f3230aa6, which
removed some unnecessary escaping of "." and "?" when printing symbols
in prin1.

This change removes more unnecessary escaping of "."; "." only
needs to be escaped when it's the entire symbol or if it's
followed by a "?".

(Actually, if we always escaped "?" (which was the case before
637dde4aba92) then "." only ever needs to be escaped when
string_to_number returns non-nil.  So 637dde4aba92 could have
just dropped the escaping of "." with no other changes, if it
didn't also remove escaping of "?")

* src/print.c (print_object): Only escape "." as the first
character in a symbol if followed by "?" or nothing (bug#23130).
* test/src/print-tests.el (test-dots): Update for new behavior.