Signal end-of-file with more correct data
end_of_file_error previously always signaled end-of-file with
load-true-file-name if that was non-nil (and a string).
However, this might be the wrong thing to do; for example, if a
file being loaded calls read on a buffer.

We can determine what's correct by also checking if readcharfun
is reading from a file.  Also, by checking that we can signal a
better error for read on a buffer.

* src/lread.c (end_of_file_error): Check readcharfun to
determine what data to signal with.  (bug#68546)
(read_char_escape, read_char_literal, read_string_literal)
(skip_space_and_comments, read0): Pass readcharfun to
end_of_file_error.