Insert * on TAB with completion-pcm-leading-wildcard=t
Improve completion-pcm-leading-wildcard by actually inserting
a * when try-completion runs.  completion-pcm--merge-try will
automatically remove the wildcard anywhere that it has no
matches, so the * is inserted only where it actually had an
effect.

This makes the behavior more explicit and consistent, and
improves performance in the case where completion-styles is
configured such that we first do partial-completion with
completion-pcm-leading-wildcard=nil, and then fall back to
completion-pcm-leading-wildcard=t: explicitly inserting the *
avoids the need to fallback.

* lisp/minibuffer.el (completion-pcm--string->pattern): Add
'star wildcard, not just 'prefix.