Display unquoted candidates in *Completions*
To fix bug#77754 and generally improve *Completions* behavior,
start displaying the unquoted versions of completion candidates
in *Completions* and in icomplete.

External completion UIs can call completion-for-display to
obtain the same behavior (and fix similar bugs in their UI).

* lisp/minibuffer.el (completion-for-display): Add.
(completion--twq-all): Add FIXME to simplify this function.
(completion--insert): Call completion-for-display. (bug#77754)
* lisp/icomplete.el (icomplete--render-vertical)
(icomplete-completions): Call completion-for-display.